Sweep preconditions to prepare for C# running
