+++ src/main/java/org/openrewrite/staticanalysis/ExplicitInitialization.java
@@ -63,7 +63,7 @@ public class ExplicitInitialization extends Recipe {
             public J visit(@Nullable Tree tree, ExecutionContext ctx) {
                 if (tree instanceof JavaSourceFile) {
                     JavaSourceFile cu = (JavaSourceFile) requireNonNull(tree);
-                    ExplicitInitializationStyle style = ((SourceFile) cu).getStyle(ExplicitInitializationStyle.class);
+                    ExplicitInitializationStyle style = cu.getStyle(ExplicitInitializationStyle.class);
                     if (style == null) {
                         style = Checkstyle.explicitInitialization();
                     }
