+++ src/main/java/org/openrewrite/staticanalysis/HideUtilityClassConstructor.java
@@ -62,7 +62,7 @@ public class HideUtilityClassConstructor extends Recipe {
         public J visit(@Nullable Tree tree, ExecutionContext ctx) {
             if (tree instanceof JavaSourceFile) {
                 JavaSourceFile cu = (JavaSourceFile) requireNonNull(tree);
-                HideUtilityClassConstructorStyle style = ((SourceFile) cu).getStyle(HideUtilityClassConstructorStyle.class);
+                HideUtilityClassConstructorStyle style = cu.getStyle(HideUtilityClassConstructorStyle.class);
                 if (style == null) {
                     style = Checkstyle.hideUtilityClassConstructorStyle();
                 }
