+++ src/main/java/org/openrewrite/staticanalysis/LambdaBlockToExpression.java
@@ -51,11 +51,12 @@ public class LambdaBlockToExpression extends Recipe {
                                 Space prefix = statement.getPrefix();
                                 if (statement instanceof J.Return) {
                                     Expression expression = ((J.Return) statement).getExpression();
-                                    if (prefix.getComments().isEmpty()) {
-                                        return l.withBody(expression);
-                                    } else {
-                                        return l.withBody(expression.withPrefix(prefix));
-                                    }
+                                        if (prefix.getComments().isEmpty()) {
+                                            //noinspection DataFlowIssue
+                                            return l.withBody(expression);
+                                        } else if (expression != null) {
+                                            return l.withBody(expression.withPrefix(prefix));
+                                        }
                                 } else if (statement instanceof J.MethodInvocation) {
                                     if (prefix.getComments().isEmpty()) {
                                         return l.withBody(statement);
