+++ src/main/java/org/openrewrite/staticanalysis/NestedEnumsAreNotStatic.java
@@ -67,6 +67,7 @@ public class NestedEnumsAreNotStatic extends Recipe {
                 if (cd.getKind() == J.ClassDeclaration.Kind.Type.Enum && cd.getType() != null && cd.getType().getOwningClass() != null) {
                     if (J.Modifier.hasModifier(cd.getModifiers(), J.Modifier.Type.Static)) {
                         J.Block enumBody = cd.getBody();
+                        //noinspection DataFlowIssue
                         cd = cd.withBody(null);
                         cd = maybeAutoFormat(cd,
                                 cd.withModifiers(ListUtils.map(cd.getModifiers(), mod ->
