+++ src/test/java/org/openrewrite/staticanalysis/ReplaceLambdaWithMethodReferenceTest.java
@@ -29,7 +29,7 @@ import org.openrewrite.test.TypeValidation;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.openrewrite.java.Assertions.java;
 
-@SuppressWarnings({"unchecked", "RedundantCast", "SimplifyStreamApiCallChains", "Convert2MethodRef", "CodeBlock2Expr", "RedundantOperationOnEmptyContainer", "ResultOfMethodCallIgnored", "rawtypes", "UnusedAssignment"})
+@SuppressWarnings({"unchecked", "RedundantCast", "SimplifyStreamApiCallChains", "Convert2MethodRef", "CodeBlock2Expr", "RedundantOperationOnEmptyContainer", "ResultOfMethodCallIgnored", "rawtypes", "UnusedAssignment", "OptionalGetWithoutIsPresent"})
 class ReplaceLambdaWithMethodReferenceTest implements RewriteTest {
 
     @Override
@@ -87,7 +87,7 @@ class ReplaceLambdaWithMethodReferenceTest implements RewriteTest {
           java(
             """
               import java.util.stream.Stream;
-              
+                            
               class Test {
                   Stream<String> method() {
                       return Stream.of(1, 32, 12, 15, 23).map(x -> Integer.toString(x));
@@ -242,7 +242,7 @@ class ReplaceLambdaWithMethodReferenceTest implements RewriteTest {
             //language=java
             """
               import org.test.CheckType;
-              
+                            
               import java.util.List;
               import java.util.stream.Collectors;
 
@@ -525,16 +525,16 @@ class ReplaceLambdaWithMethodReferenceTest implements RewriteTest {
     @Test
     void castType() {
         rewriteRun(
-          //language=java
           java(
+            //language=java
             """
               package org.test;
               public class CheckType {
               }
               """
           ),
-          //language=java
           java(
+            //language=java
             """
               import java.util.List;
               import java.util.stream.Collectors;
@@ -550,6 +550,7 @@ class ReplaceLambdaWithMethodReferenceTest implements RewriteTest {
                   }
               }
               """,
+            //language=java
             """
               import java.util.List;
               import java.util.stream.Collectors;
@@ -606,7 +607,7 @@ class ReplaceLambdaWithMethodReferenceTest implements RewriteTest {
             //language=java
             """
               import org.test.CheckType;
-              
+                            
               import java.util.List;
               import java.util.stream.Collectors;
 
@@ -1335,21 +1336,21 @@ class ReplaceLambdaWithMethodReferenceTest implements RewriteTest {
           //language=java
           java(
             """
-            import java.util.*;
-            import java.util.stream.*;
-            
-            class Animal {}
-            class Cat extends Animal {}
-            class Dog extends Animal {}
+              import java.util.*;
+              import java.util.stream.*;
+                          
+              class Animal {}
+              class Cat extends Animal {}
+              class Dog extends Animal {}
 
-            class Test {
-              public void groupOnGetClass() {
-                List<Animal> animals = List.of(new Cat(), new Dog());
-                Map<Class<? extends Animal>, List<Animal>> collect;
-                collect = animals.stream().collect(Collectors.groupingBy(a -> a.getClass()));
+              class Test {
+                public void groupOnGetClass() {
+                  List<Animal> animals = List.of(new Cat(), new Dog());
+                  Map<Class<? extends Animal>, List<Animal>> collect;
+                  collect = animals.stream().collect(Collectors.groupingBy(a -> a.getClass()));
+                }
               }
-            }
-            """
+              """
           )
         );
     }
